+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### [Fixed]
 
+* Dataset : debug pour les dossiers non relatif au fichier descripteur [#260](https://github.com/Geoplateforme/sdk-entrepot/issues/260)
+
 ## v0.1.42
 
 ### [Added]