+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### [Fixed]
 
-* Dataset : debug pour les dossiers non relatif au fichier descripteur [#260](https://github.com/Geoplateforme/sdk-entrepot/issues/260)
+* Dataset : debug pour les dossiers non relatifs au fichier descripteur [#260](https://github.com/Geoplateforme/sdk-entrepot/issues/260)
 
 ## v0.1.42